feat: update and show disclaimer
