+++ tasks/.env.production
@@ -3,3 +3,4 @@ NEXT_PUBLIC_API_URL=https://staging.api.helpwave.de
 NEXT_PUBLIC_MOCK=true
 NEXT_PUBLIC_OAUTH_REDIRECT_URI=https://staging-tasks.helpwave.de/auth/callback
 NEXT_PUBLIC_FAKE_TOKEN_ENABLE=true
+NEXT_PUBLIC_SHOW_STAGING_DISCLAIMER_MODAL=true
